Allow to pass session cookies as command line argument
Some users may have to use SAML and multi-factor authentication (MFA).
Having them to pass in a session cookie (SVPNCOOKIE) allows to move
the responsibility of interfacing with third party systems to some
external program and have openfortivpn skip the authentication phase.

The cookie must be in the form `SVPNCOOKIE=cookie_value`.